feat: update and delete category

+++ src/application/useCases/UpdateCategory.ts
@@ -1,5 +1,6 @@
 import { CategogyRepository } from "application/repositories/CategoryRepository";
 import { Category } from "domain/entities/Category";
+import { CategoryNotFound } from "domain/errors/CategoryNotFound";
 import { Storage } from "infra/storage/Storage";
 
 export class UpdateCategory{
@@ -11,6 +12,10 @@ export class UpdateCategory{
     async execute(id:string,name:string, file?:string):Promise<Category>{
         const category=await this.categoryRepository.findById(id)
         
+        if(category===null){
+            throw new CategoryNotFound();
+        }
+
         if(file !== undefined){
             this.deleteOldIcon(category)
         }
@@ -23,6 +28,8 @@ export class UpdateCategory{
     }
 
     private deleteOldIcon(category){
-        this.storage.deleteFile(category.icon)
+        if(category.icon){
+            this.storage.deleteFile(category.icon)
+        }
     }
 }
